Fixed map intent when there is no geo location for a track.  Clicking on a track previously would of opened maps with a null location.  Now a toast makes the user aware there is no GPS information stored with the track.
Fixed the add track day dialog popping up when no tracks are in the track list.  Previously the dialog would appear and be empty.  Now a toast makes the user aware there are no tracks in the track list.

Added a check in the DetailsTextWatcher that makes sure the context passed to it implements the callback necessary for it to function.  Now it will throw a ClassCastException if it does not.